Added docker studio support to all studio subcommands
