+++ components/hab/src/command/pkg/build.rs
@@ -27,6 +27,7 @@ pub fn start(
     keys: Option<&str>,
     reuse: bool,
     windows: bool,
+    docker: bool,
 ) -> Result<()> {
     let mut args: Vec<OsString> = Vec::new();
     if let Some(root) = root {
@@ -49,5 +50,8 @@ pub fn start(
     if cfg!(target_os = "windows") && windows {
         args.push("-w".into());
     }
-    studio::start(ui, args)
+    if cfg!(target_os = "linux") && docker {
+        args.push("-D".into());
+    }
+    studio::enter::start(ui, args)
 }
