+++ components/hab/src/command/studio/docker.rs
@@ -12,156 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::env;
-use std::fs as stdfs;
-use std::ffi::OsString;
-use std::path::PathBuf;
-
-use common::ui::UI;
-use hcore::crypto::CACHE_KEY_PATH_ENV_VAR;
-use hcore::env as henv;
-use hcore::fs;
-
-use config;
-use error::Result;
-
-const ARTIFACT_PATH_ENVVAR: &'static str = "ARTIFACT_PATH";
-const ORIGIN_ENVVAR: &'static str = "HAB_ORIGIN";
-const STUDIO_CMD: &'static str = "hab-studio";
-const STUDIO_CMD_ENVVAR: &'static str = "HAB_STUDIO_BINARY";
-const STUDIO_PACKAGE_IDENT: &'static str = "core/hab-studio";
-
-pub fn start(ui: &mut UI, args: Vec<OsString>) -> Result<()> {
-    if henv::var(ORIGIN_ENVVAR).is_err() {
-        let config = config::load()?;
-        if let Some(default_origin) = config.origin {
-            debug!("Setting default origin {} via CLI config", &default_origin);
-            env::set_var("HAB_ORIGIN", default_origin);
-        }
-    }
-
-    if henv::var(CACHE_KEY_PATH_ENV_VAR).is_err() {
-        let path = fs::cache_key_path(None::<&str>);
-        debug!("Setting {}={}", CACHE_KEY_PATH_ENV_VAR, path.display());
-        env::set_var(CACHE_KEY_PATH_ENV_VAR, &path);
-    };
-
-    let artifact_path = match henv::var(ARTIFACT_PATH_ENVVAR) {
-        Ok(p) => PathBuf::from(p),
-        Err(_) => {
-            let path = fs::cache_artifact_path(None::<&str>);
-            debug!("Setting {}={}", ARTIFACT_PATH_ENVVAR, path.display());
-            env::set_var(ARTIFACT_PATH_ENVVAR, &path);
-            path
-        }
-    };
-    if !artifact_path.is_dir() {
-        debug!("Creating artifact_path at: {}", artifact_path.display());
-        stdfs::create_dir_all(&artifact_path)?;
-    }
-
-    inner::start(ui, args)
-}
-
-#[cfg(target_os = "linux")]
-mod inner {
-    use std::env;
-    use std::ffi::OsString;
-    use std::path::PathBuf;
-    use std::str::FromStr;
-
-    use common::ui::UI;
-    use hcore::crypto::{init, default_cache_key_path};
-    use hcore::env as henv;
-    use hcore::fs::{am_i_root, find_command};
-    use hcore::os::process;
-    use hcore::package::PackageIdent;
-
-    use error::{Error, Result};
-    use exec;
-    use VERSION;
-
-    const SUDO_CMD: &'static str = "sudo";
-
-    pub fn start(ui: &mut UI, args: Vec<OsString>) -> Result<()> {
-        rerun_with_sudo_if_needed(ui)?;
-
-        let command = match henv::var(super::STUDIO_CMD_ENVVAR) {
-            Ok(command) => PathBuf::from(command),
-            Err(_) => {
-                init();
-                let version: Vec<&str> = VERSION.split("/").collect();
-                let ident = PackageIdent::from_str(
-                    &format!("{}/{}", super::STUDIO_PACKAGE_IDENT, version[0]),
-                )?;
-                exec::command_from_min_pkg(
-                    ui,
-                    super::STUDIO_CMD,
-                    &ident,
-                    &default_cache_key_path(None),
-                    0,
-                )?
-            }
-        };
-
-        if let Some(cmd) = find_command(command.to_string_lossy().as_ref()) {
-            Ok(process::become_command(cmd, args)?)
-        } else {
-            Err(Error::ExecCommandNotFound(command))
-        }
-    }
-
-    fn rerun_with_sudo_if_needed(ui: &mut UI) -> Result<()> {
-        // If I have root permissions, early return, we are done.
-        if am_i_root() {
-            return Ok(());
-        }
-
-        // Otherwise we will try to re-run this program using `sudo`
-        match find_command(SUDO_CMD) {
-            Some(sudo_prog) => {
-                let mut args: Vec<OsString> = vec![
-                    "-p".into(),
-                    "[sudo hab-studio] password for %u: ".into(),
-                    "-E".into(),
-                ];
-                args.append(&mut env::args_os().collect());
-                Ok(process::become_command(sudo_prog, args)?)
-            }
-            None => {
-                ui.warn(format!(
-                    "Could not find the `{}' command, is it in your PATH?",
-                    SUDO_CMD
-                ))?;
-                ui.warn(
-                    "Running Habitat Studio requires root or administrator privileges. \
-                              Please retry this command as a super user or use a \
-                              privilege-granting facility such as sudo.",
-                )?;
-                ui.br()?;
-                Err(Error::RootRequired)
-            }
-        }
-    }
-}
-
-#[cfg(not(target_os = "linux"))]
-mod inner {
     use std::env;
     use std::ffi::{OsStr, OsString};
     use std::path::{Path, PathBuf};
     use std::process::{Command, Stdio};
-    use std::str::FromStr;
 
     use common::ui::UI;
-    use hcore::crypto::{init, default_cache_key_path};
+    use hcore::crypto::{default_cache_key_path};
     use hcore::env as henv;
     use hcore::fs::{CACHE_ARTIFACT_PATH, CACHE_KEY_PATH, find_command};
     use hcore::os::process;
-    use hcore::package::PackageIdent;
+    use command::studio::enter::ARTIFACT_PATH_ENVVAR;
 
     use error::{Error, Result};
-    use exec;
     use VERSION;
 
     const DOCKER_CMD: &'static str = "docker";
@@ -171,42 +34,7 @@ mod inner {
     const DOCKER_OPTS_ENVVAR: &'static str = "HAB_DOCKER_OPTS";
     const DOCKER_SOCKET: &'static str = "/var/run/docker.sock";
 
-    pub fn start(_ui: &mut UI, args: Vec<OsString>) -> Result<()> {
-        if is_windows_studio(&args) {
-            start_windows_studio(_ui, args)
-        } else {
-            start_docker_studio(_ui, args)
-        }
-    }
-
-    pub fn start_windows_studio(ui: &mut UI, args: Vec<OsString>) -> Result<()> {
-        let command = match henv::var(super::STUDIO_CMD_ENVVAR) {
-            Ok(command) => PathBuf::from(command),
-            Err(_) => {
-                init();
-                let version: Vec<&str> = VERSION.split("/").collect();
-                let ident = PackageIdent::from_str(
-                    &format!("{}/{}", super::STUDIO_PACKAGE_IDENT, version[0]),
-                )?;
-                exec::command_from_min_pkg(
-                    ui,
-                    super::STUDIO_CMD,
-                    &ident,
-                    &default_cache_key_path(None),
-                    0,
-                )?
-            }
-        };
-
-        if let Some(cmd) = find_command(command.to_string_lossy().as_ref()) {
-            process::become_command(cmd, args)?;
-        } else {
-            return Err(Error::ExecCommandNotFound(command));
-        }
-        Ok(())
-    }
-
-    pub fn start_docker_studio(_ui: &mut UI, args: Vec<OsString>) -> Result<()> {
+    pub fn start_docker_studio(_ui: &mut UI, mut args: Vec<OsString>) -> Result<()> {
         let docker_cmd = find_docker_cmd()?;
 
         if is_image_present(&docker_cmd) {
@@ -224,7 +52,7 @@ mod inner {
                 CACHE_KEY_PATH
             ),
         ];
-        if let Ok(cache_artifact_path) = henv::var(super::ARTIFACT_PATH_ENVVAR) {
+        if let Ok(cache_artifact_path) = henv::var(ARTIFACT_PATH_ENVVAR) {
             volumes.push(format!("{}:/{}", cache_artifact_path, CACHE_ARTIFACT_PATH));
         }
         if Path::new(DOCKER_SOCKET).exists() || cfg!(target_os = "windows") {
@@ -240,10 +68,16 @@ mod inner {
             "http_proxy",
             "https_proxy",
         ];
+        // We need to strip out the -D if it exists to avoid
+        // it getting passed to the sup on entering the studio
+        let to_cull = OsString::from("-D");
+        let index = args.iter().position(|x| *x == to_cull).unwrap();
+        args.remove(index);
 
         check_mounts(&docker_cmd, volumes.iter())?;
         run_container(docker_cmd, args, volumes.iter(), env_vars.iter())
     }
+
 
     fn find_docker_cmd() -> Result<PathBuf> {
         let docker_cmd = henv::var(DOCKER_CMD_ENVVAR).unwrap_or(DOCKER_CMD.to_string());
@@ -403,30 +237,3 @@ mod inner {
         let version: Vec<&str> = VERSION.split("/").collect();
         henv::var(DOCKER_IMAGE_ENVVAR).unwrap_or(format!("{}:{}", DOCKER_IMAGE, version[0]))
     }
-
-    fn is_windows_studio(args: &Vec<OsString>) -> bool {
-        if cfg!(not(target_os = "windows")) {
-            return false;
-        }
-
-        for arg in args.iter() {
-            let str_arg = arg.to_string_lossy().to_lowercase();
-            if str_arg == String::from("--windows") || str_arg == String::from("-w") {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    #[cfg(test)]
-    mod tests {
-        use super::{image_identifier, DOCKER_IMAGE};
-        use VERSION;
-
-        #[test]
-        fn retrieve_image_identifier() {
-            assert_eq!(image_identifier(), format!("{}:{}", DOCKER_IMAGE, VERSION));
-        }
-    }
-}
